Fix ReferringSitesPage Title Height & Improve Performance
