+++ GitTrends.Android/Properties/AndroidManifest.xml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" android:versionCode="32" android:versionName="1.2.2" package="com.minnick.gittrends" android:installLocation="auto">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" android:versionCode="33" android:versionName="1.2.3" package="com.minnick.gittrends" android:installLocation="auto">
 	<uses-sdk android:minSdkVersion="21" android:targetSdkVersion="29" />
 	<uses-permission android:name="com.google.android.c2dm.permission.RECEIVE" />
 	<application android:label="GitTrends" android:fullBackupContent="@xml/auto_backup_rules">
